integrated global-style components to sharedModule to be available to all feature modules
