+++ src/app/app-routing.module.ts
@@ -1,4 +1,4 @@
-git import { NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 const routes: Routes = [
