+++ src/app/user/user.module.ts
@@ -1,17 +1,16 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { UserRoutingModule } from './user-routing.module';
+import { SharedModule } from '../global-components/shared.module';
 
 
 
 @NgModule({
-  declarations: [/* Main component goes here */],
-  imports: [
-    CommonModule,
-    UserRoutingModule
-  ],
-  exports: [
+  declarations: [
     /* Main component goes here */
+  ],
+  imports: [
+    SharedModule,
+    UserRoutingModule
   ]
 })
 export class UserModule { }
